+++ packages/ai-genkit/package.json
@@ -3,6 +3,11 @@
   "version": "0.0.0",
   "description": "",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "module": "./dist/index.mjs",
+  "files": [
+    "dist/**/*"
+  ],
   "scripts": {
     "build": "tsc",
     "doc": "typedoc src/index.ts",
