+++ packages/ai-genkit/tsconfig.json
@@ -1,12 +1,26 @@
 {
   "compilerOptions": {
-    "module": "CommonJS",
-    "target": "es2015",
+    "composite": false,
     "declaration": true,
-    "outDir": "./dist",
+    "declarationMap": true,
     "esModuleInterop": true,
-    "skipLibCheck": true
+    "forceConsistentCasingInFileNames": true,
+    "inlineSources": false,
+    "isolatedModules": true,
+    "moduleResolution": "node",
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "preserveWatchOutput": true,
+    "skipLibCheck": true,
+    "strict": true,
+    "types": ["@types/node", "vitest/globals"],
+    "lib": ["dom", "ES2018"],
+    "module": "ESNext",
+    "target": "ES2018",
+    "outDir": "./dist",
+    "baseUrl": "."
   },
   "include": ["src/**/*"],
-  "types": ["@types/node"]
+  "types": ["@types/node"],
+  "exclude": ["node_modules"]
 }
